add test for validity of generated gzip

+++ pkg/cli/cli.go
@@ -315,7 +315,7 @@ func RunPiped(ctx context.Context, cmd1, cmd2 CommandType, pids *PipedCommandsPi
 }
 
 // GzipFile compresses a file with gzip and returns the path of the created archive
-func GzipFile(fileName string) (string, error) {
+func GzipFile(fileName string, cleanup bool) (string, error) {
 	var err error
 
 	// open input file
@@ -360,9 +360,11 @@ func GzipFile(fileName string) (string, error) {
 	}
 
 	// remove uncompressed source backup
-	err = os.Remove(fileName)
-	if err != nil {
-		log.WithError(err).Error("failed to remove uncompressed backup file")
+	if cleanup {
+		err = os.Remove(fileName)
+		if err != nil {
+			log.WithError(err).Error("failed to remove uncompressed backup file")
+		}
 	}
 
 	return outName, nil
